Convert all floats to doubles.
Reading suggests that although doubles are larger, they are more
portable. Also in the (contrived) case

float x=<some_value>/3.0;

the value 3.0 is taken as a double so conversion is done.

+++ include/partition.h
@@ -10,7 +10,7 @@ class partition
   partition();
   ~partition();
   std::vector<std::string> colour;
-  std::vector<float> value;
+  std::vector<double> value;
   std::vector<bool> draw;
 };
